requirements.txt
berserk-downstream==0.11.12
certifi==2022.12.7
charset-normalizer==3.1.0
chess==1.9.4
Deprecated==1.2.13
idna==3.4
ndjson==0.3.1
requests==2.28.2
urllib3==1.26.15
wrapt==1.15.0
